backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
flake8==7.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
packaging==24.2
pluggy==1.5.0
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.8.0
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
